Rename AxiosHttpClient to HttpClient. Remove request from package.json Fix HttpClient unit test failures.
