+++ broker/applications/quota-app/package.json
@@ -19,7 +19,7 @@
     "@sf/logger": "file:../../core/logging",
     "@sf/quota": "file:../../data-access-layer/quota",
     "ajv": "6.12.6",
-    "axios": "^0.20.0",
+    "axios": "0.20.0",
     "basic-auth": "2.0.1",
     "bluebird": "3.7.2",
     "body-parser": "1.19.0",
@@ -40,7 +40,6 @@
     "morgan": "1.10.0",
     "pubsub-js": "1.8.0",
     "pug": "3.0.0",
-    "request": "2.88.2",
     "riemann": "2.0.0",
     "rxjs": "5.0.0",
     "uuid": "8.2.0",
