+++ broker/applications/reports/package.json
@@ -24,7 +24,7 @@
     "ajv": "6.12.6",
     "ali-oss": "6.10.0",
     "aws-sdk": "2.747.0",
-    "axios": "^0.20.0",
+    "axios": "0.20.0",
     "azure-arm-compute": "10.0.0",
     "azure-storage": "2.10.3",
     "basic-auth": "2.0.1",
@@ -51,7 +51,6 @@
     "pkgcloud": "1.7.0",
     "pubsub-js": "1.8.0",
     "pug": "3.0.0",
-    "request": "2.88.2",
     "riemann": "2.0.0",
     "rxjs": "5.0.0",
     "session-file-store": "1.4.0",
