+++ broker/core/utils/src/index.js
@@ -4,7 +4,6 @@ const commonFunctions = require('./commonFunctions');
 const commonVarliables = require('./commonVariables');
 const serviceFlowMapper = require('./ServiceFlowMapper');
 const HttpClient = require('./HttpClient');
-const AxiosHttpClient = require('./AxiosHttpClient');
 const errors = require('./errors');
 const JWT = require('./jwt');
 const Repository = require('./Repository');
@@ -18,7 +17,6 @@ module.exports = {
   CONST: commonVarliables,
   serviceFlowMapper,
   HttpClient,
-  AxiosHttpClient,
   JWT,
   Repository,
   EncryptionManager,
