+++ broker/data-access-layer/bosh/test/bosh.BoshDirectorClient.spec.js
@@ -12,7 +12,7 @@ const {
     DirectorServiceUnavailable,
     UnprocessableEntity
   },
-  AxiosHttpClient,
+  HttpClient,
   EncryptionManager
 } = require('@sf/common-utils');
 const config = require('@sf/app-config');
@@ -1321,7 +1321,7 @@ describe('bosh', () => {
         };
         let tokenNotExpired = 'eyJhbGciOiJIUzI1NiJ9.eyJleHAiOjM4MzQ4NjQwMDB9';
         let sandbox = sinon.createSandbox();
-        let requestStub = sandbox.stub(AxiosHttpClient.prototype, 'request');
+        let requestStub = sandbox.stub(HttpClient.prototype, 'request');
         // create actual boshDirectorClient
         let dummyBoshDirectorClient = new BoshDirectorClient();
         let getAccessTokenBoshUAAStub = sandbox.stub(dummyBoshDirectorClient.uaaObjects[directorConfig.name].tokenIssuer, 'getAccessTokenBoshUAA');
